requirements.txt
Django==1.9.7
django-debug-toolbar==1.4
MySQL-python==1.2.5
sqlparse==0.2.1
mysql-connector==2.1.6
grpcio==1.17.0
grpcio-tools==1.17.0
google-compute-engine==2.8.3
git+https://github.com/Benjamin-Dewey/pickledb
